+++ recreate_nextcloud_db.sh
@@ -13,21 +13,37 @@ set -u
 set -e
 
 source './config.env'
+source './functions.sh'
 
-echo "Dropping existing database '${MYSQL_DATABASE}' (if it exists)..."
-mysql \
-    --host="${MYSQL_HOST}" \
-    --user="${MYSQL_USER}" \
-    --password="${MYSQL_PASSWORD}" \
-    --execute="DROP DATABASE IF EXISTS ${MYSQL_DATABASE}" \
-    --verbose
-echo ""
+if [[ "${DELETE_PROMPT:-1}" -eq 1 ]]; then
+    {
+        echo "This will attempt to remove the Nextcloud database, which will"
+        echo "result in the loss of Nextcloud configuration information, user"
+        echo "information, audit logs, etc."
+        echo ""
+    } >&2
 
-echo "Re-creating database '${MYSQL_DATABASE}'..."
-mysql \
-    --host="${MYSQL_HOST}" \
-    --user="${MYSQL_USER}" \
-    --password="${MYSQL_PASSWORD}" \
-    --execute="CREATE DATABASE ${MYSQL_DATABASE}" \
-    --verbose
-echo ""
+    confirmation_prompt "Are you sure"
+else
+    confirmed=1
+fi
+
+if [[ "${confirmed}" -eq 1 ]]; then
+    echo "Dropping existing database '${MYSQL_DATABASE}' (if it exists)..."
+    mysql \
+        --host="${MYSQL_HOST}" \
+        --user="${MYSQL_USER}" \
+        --password="${MYSQL_PASSWORD}" \
+        --execute="DROP DATABASE IF EXISTS ${MYSQL_DATABASE}" \
+        --verbose
+    echo ""
+
+    echo "Re-creating database '${MYSQL_DATABASE}'..."
+    mysql \
+        --host="${MYSQL_HOST}" \
+        --user="${MYSQL_USER}" \
+        --password="${MYSQL_PASSWORD}" \
+        --execute="CREATE DATABASE ${MYSQL_DATABASE}" \
+        --verbose
+    echo ""
+fi
